gsl: skip unsuccessful PD service registration for a subsystem
On Kaanapali Qultivate SKU2 variant, PD notifier registration for CDSP service is failing because CDSP is not present on this variant.
This, in turn, is causing GSL initialization failure.

Ideally, the servreg domain list could have indicated this, but it may not be possible due to cross-platform build dependencies.
Therefore, ignore the error to prevent GSL initialization failure.

Change-Id: Iddd0016e149fc21667a63f2bdd8ed09809cac373